committing

+++ temperature/models.py
@@ -8,7 +8,7 @@ class location(models.Model):
     access_key=models.CharField(max_length=200)
 
 class tempEntry(models.Model):
-    temp=models.IntegerField()
+    temp=models.FloatField()
     time=models.IntegerField()
     place=models.ForeignKey(location, on_delete=models.CASCADE, default=1)
